fullscreen/api/historical.html

<!doctype html>
<title>Historical Fullscreen features</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
[
  "webkitRequestFullScreen",
  "onwebkitfullscreenchange",
  "onwebkitfullscreenerror",
].forEach(function(member) {
  test(function() {
    assert_false(member in document.createElement('div'));
  }, 'Element member must be nuked: ' + member);
});
</script>
